updating queuing

+++ preprocess_service/code/basic_preprocess.py
@@ -3,12 +3,14 @@ Simple test for celery to run preprocess tasks
 """
 import sys
 from datetime import datetime
+import time
 import json
 from os.path import \
-    (abspath, basename, dirname, isdir, join, splitext)
-from celery import Celery
-
+    (abspath, basename, dirname, isdir, isfile, join, splitext)
 import logging
+
+from celery import Celery
+from random_util import get_alphanumeric_lowercase
 
 # Add preprocess code dir
 PREPROCESS_DIR = join(dirname(dirname(dirname(abspath(__file__)))),
@@ -32,9 +34,12 @@ app = Celery('basic_preprocess',
 def preprocess_csv_file(input_file, output_dir):
     """Run preprocess on a csv file"""
     init_timestamp = datetime.now()
+    start_time = time.time()
+
     print('(%s) Start preprocess: %s' % (init_timestamp, input_file))
     if not isdir(output_dir):
         return dict(success=False,
+                    input_file=input_file,
                     message='Directory does not exist: %s' % output_dir)
 
     # Split out the filename and extension
@@ -49,8 +54,9 @@ def preprocess_csv_file(input_file, output_dir):
         runner, user_msg = PreprocessRunner.load_from_csv_file(input_file)
 
     if user_msg:
-        print('(%s) FAILED: %s' % (init_timestamp, user_msg))
+        print('(%s) FAILED: %s' % (input_file, user_msg))
         return dict(success=False,
+                    input_file=input_file,
                     message=user_msg)
 
 
@@ -61,20 +67,46 @@ def preprocess_csv_file(input_file, output_dir):
     # write to file
     # ------------------------------
     jstring = runner.get_final_json(indent=4)
-    output_filepath = join(output_dir, '%s.json' % fname_base)
+
+    # create output filename
+    output_filepath = None
+    for _idx in range(5):
+        output_filepath = '%s_%s.json' % \
+                          (fname_base, get_alphanumeric_lowercase(4))
+        output_filepath = join(output_dir, output_filepath)
+        if not isfile(output_filepath):
+            break
+
+    if isfile(output_filepath):
+        user_msg = 'Output file already exists (tried 5 random names): %s' % \
+                    (output_filepath)
+        print('%s FAILED: %s' % (input_file, user_msg))
+        return dict(success=False,
+                    input_file=input_file,
+                    message=user_msg)
 
     try:
         open(output_filepath, 'w').write(jstring)
     except OSError as os_err:
         user_msg = 'Failed to write file: %s' % (os_err)
-        print('(%s) FAILED: %s' % (init_timestamp, user_msg))
+        print('(%s) FAILED: %s' % (input_file, user_msg))
         return dict(success=False,
+                    input_file=input_file,
                     message=user_msg)
 
     user_msg = 'file written: %s' % output_filepath
     print('(%s) SUCCESS: %s' % (init_timestamp, user_msg))
 
+    # your script
+    elapsed_time = time.time() - start_time
+    elapsed_time_str = time.strftime("%H:%M:%S", time.gmtime(elapsed_time))
+
+    print('   - elapsed time: %s' % (elapsed_time_str))
+
+
     return dict(success=True,
+                input_file=input_file,
                 message=user_msg,
+                elapsed_time=elapsed_time_str,
                 data=runner.get_final_json())
                 #data=runner.get_final_dict())
